// net/src/achievements/catalog.gen.ts
export interface CatalogEntry {
    id: string;
    steam: string;
    hidden: boolean;
    minHours: number;
    minSessionMinutes: number;
    requires: string[];
}